+++ cmd/netassert/cli/ping.go
@@ -2,35 +2,50 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"os"
+	"time"
 
 	"github.com/spf13/cobra"
+
+	"github.com/controlplaneio/netassert/v2/internal/logger"
 )
 
-// health endpoint for the K8s server
-var apiServerHealthEndpoint = `/healthz`
+const (
+	apiServerHealthEndpoint = `/healthz` // health endpoint for the K8s server
+)
+
+type pingCmdConfig struct {
+	KubeConfig  string
+	PingTimeout time.Duration
+}
+
+var pingCmdCfg = pingCmdConfig{}
 
 var pingCmd = &cobra.Command{
 	Use: "ping",
-	Short: "pings the K8s API server over HTTP to see if it is alive and also checks if the server has support for " +
-		"ephemeral containers",
-	Long: "pings the K8s API server over HTTP to see if it is alive and also checks if the server has support for " +
-		"ephemeral/debug containers",
+	Short: "pings the K8s API server over HTTP(S) to see if it is alive and also checks if the server has support for " +
+		"ephemeral containers.",
+	Long: "pings the K8s API server over HTTP(S) to see if it is alive and also checks if the server has support for " +
+		"ephemeral/debug containers.",
 	Run: func(cmd *cobra.Command, args []string) {
-		ping()
+		ctx, cancel := context.WithTimeout(context.Background(), pingCmdCfg.PingTimeout)
+		defer cancel()
+		ping(ctx)
 	},
 	Version: rootCmd.Version,
 }
 
 // checkEphemeralContainerSupport checks to see if ephemeral containers are supported by the K8s server
-func ping() {
-	k8sSvc, err := createService()
+func ping(ctx context.Context) {
+	lg := logger.NewHCLogger("info", fmt.Sprintf("%s-%s", appName, version), os.Stdout)
+
+	k8sSvc, err := createService(pingCmdCfg.KubeConfig, lg)
+
 	if err != nil {
-		lg.Error("Failed to build K8s Client", "error", err)
+		lg.Error("Ping failed, unable to build K8s Client", "error", err)
 		os.Exit(1)
 	}
-
-	ctx := context.Background()
 
 	if err := k8sSvc.PingHealthEndpoint(ctx, apiServerHealthEndpoint); err != nil {
 		lg.Error("Ping failed", "error", err)
@@ -46,4 +61,10 @@ func ping() {
 	}
 
 	lg.Info("✅ Ephemeral containers are supported by the Kubernetes server")
+}
+
+func init() {
+	pingCmd.Flags().DurationVarP(&pingCmdCfg.PingTimeout, "timeout", "t", 60*time.Second,
+		"Timeout for the ping command")
+	pingCmd.Flags().StringVarP(&pingCmdCfg.KubeConfig, "kubeconfig", "k", "", "path to kubeconfig file")
 }