+++ e2e/clusters/aws-eks-terraform-module/eks.tf
@@ -8,7 +8,8 @@ data "aws_availability_zones" "available" {}
 
 module "eks" {
   source  = "terraform-aws-modules/eks/aws"
-  version = "19.5.1"
+  version = "~> 19.0"
+
 
   cluster_name    = var.cluster_name
   cluster_version = var.cluster_version