+++ internal/engine/engine_deployment_test.go
@@ -5,13 +5,13 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/controlplaneio/netassert/v2/internal/data"
 	"github.com/hashicorp/go-hclog"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"go.uber.org/mock/gomock"
 
-	"github.com/golang/mock/gomock"
+	"github.com/controlplaneio/netassert/v2/internal/data"
 )
 
 func TestEngine_GetPod_Deployment(t *testing.T) {