+++ CHANGELOG.md
@@ -4,12 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## Table of Contents
 
+- [2.0.3](#203)
 - [2.0.2](#202)
 - [2.0.1](#201)
 - [2.0.0](#200)
 - [0.1.0](#010)
 
 ---
+
+## `2.0.3`
+
+- test with latest version of Kubernetes and update to Go 1.21
+- update e2e tests with latest version of EKS and GKE and Calico CNI
 
 ## `2.0.2`